moved to category. cleaned code
