+++ conf/db.js.example
@@ -3,6 +3,6 @@ sql = {
     user     : 'USER',
     password : 'PASSWORD',
     database : 'DATABASE'
-}
+};
 
 module.exports = sql;
